Avances en panelBuscar

+++ src/main/java/pos/logic/Producto.java
@@ -45,7 +45,7 @@ public class Producto {
         return precioUnitario;
     }
 
-    public float getExistencias() {
+    public int getExistencias() {
         return existencias;
     }
